+++ lib/octantis_web/live/components/polaris/empty_state.ex
@@ -15,7 +15,7 @@ defmodule OctantisWeb.Components.Polaris.EmptyState do
    - https://github.com/Shopify/polaris/blob/main/polaris-react/src/components/EmptyState/EmptyState.tsx
   """
 
-  use OctantisWeb.Core
+  use OctantisWeb.Core, :component
 
   require OctantisWeb.Components.Polaris.ComplexAction