+++ lib/octantis_web/live/components/polaris/thumbnail.ex
@@ -12,7 +12,7 @@ defmodule OctantisWeb.Components.Polaris.Thumbnail do
    - https://github.com/Shopify/polaris/blob/main/polaris-react/src/components/Thumbnail/Thumbnail.tsx
   """
 
-  use OctantisWeb.Core
+  use OctantisWeb.Core, :component
 
   import_polaris_components([:image])